Use JTS to fix the xthreshold clipping. See http://openmap.bbn.com/mailArchives/openmap-users/2012-03/6867.html for more details.

+++ src/openmap/com/bbn/openmap/proj/coords/LatLonGCT.java
@@ -44,7 +44,7 @@ public class LatLonGCT implements GeoCoordTransformation {
     }
 
     public LatLonPoint inverse(double x, double y, LatLonPoint ret) {
-        ret.setLatLon((float) y, (float) x);
+        ret.setLatLon(y, x);
         return ret;
     }
